+++ frontend/src/metabase/querying/filters/utils.ts
@@ -1,17 +1,6 @@
 import * as Lib from "metabase-lib";
 
-import {
-  DATE_PICKER_EXTRACTION_UNITS,
-  DATE_PICKER_OPERATORS,
-  DATE_PICKER_TRUNCATION_UNITS,
-} from "./constants";
-import type {
-  DatePickerExtractionUnit,
-  DatePickerOperator,
-  DatePickerTruncationUnit,
-  DatePickerUnit,
-  FilterOperatorOption,
-} from "./types";
+import type { FilterOperatorOption } from "./types";
 
 export function getGroupName(
   groupInfo: Lib.ColumnGroupDisplayInfo,
@@ -20,31 +9,6 @@ export function getGroupName(
   return groupInfo.isMainGroup && stageIndex > 1
     ? `${groupInfo.displayName} (${stageIndex})`
     : groupInfo.displayName;
-}
-
-export function isDatePickerOperator(
-  operator: string,
-): operator is DatePickerOperator {
-  const operators: ReadonlyArray<string> = DATE_PICKER_OPERATORS;
-  return operators.includes(operator);
-}
-
-export function isDatePickerUnit(unit: string): unit is DatePickerUnit {
-  return isDatePickerTruncationUnit(unit) || isDatePickerExtractionUnit(unit);
-}
-
-export function isDatePickerTruncationUnit(
-  unit: string,
-): unit is DatePickerTruncationUnit {
-  const units: ReadonlyArray<string> = DATE_PICKER_TRUNCATION_UNITS;
-  return units.includes(unit);
-}
-
-export function isDatePickerExtractionUnit(
-  unit: string,
-): unit is DatePickerExtractionUnit {
-  const units: ReadonlyArray<string> = DATE_PICKER_EXTRACTION_UNITS;
-  return units.includes(unit);
 }
 
 export function getAvailableOperatorOptions<
