+++ frontend/src/metabase/querying/filters/utils/dates.ts
@@ -1,0 +1,36 @@
+import {
+  DATE_PICKER_EXTRACTION_UNITS,
+  DATE_PICKER_OPERATORS,
+  DATE_PICKER_TRUNCATION_UNITS,
+} from "metabase/querying/filters/constants";
+import type {
+  DatePickerExtractionUnit,
+  DatePickerOperator,
+  DatePickerTruncationUnit,
+  DatePickerUnit,
+} from "metabase/querying/filters/types";
+
+export function isDatePickerOperator(
+  operator: string,
+): operator is DatePickerOperator {
+  const operators: ReadonlyArray<string> = DATE_PICKER_OPERATORS;
+  return operators.includes(operator);
+}
+
+export function isDatePickerUnit(unit: string): unit is DatePickerUnit {
+  return isDatePickerTruncationUnit(unit) || isDatePickerExtractionUnit(unit);
+}
+
+export function isDatePickerTruncationUnit(
+  unit: string,
+): unit is DatePickerTruncationUnit {
+  const units: ReadonlyArray<string> = DATE_PICKER_TRUNCATION_UNITS;
+  return units.includes(unit);
+}
+
+export function isDatePickerExtractionUnit(
+  unit: string,
+): unit is DatePickerExtractionUnit {
+  const units: ReadonlyArray<string> = DATE_PICKER_EXTRACTION_UNITS;
+  return units.includes(unit);
+}
